requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch3d>=0.7.5
ninja>=1.11.0
click>=8.1.4
pytorch_lightning>=2.1.0
six>=1.16.0
open3d>=0.17.0
hdbscan>=0.8.28
natsort>=8.1.0
pykeops>=2.1.2
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.25.0
einops>=0.7.0
torchmetrics>=1.2.0
